+++ internal/resource/config/oauth/authserversettings/scopes/commonscope/oauth_auth_server_settings_scopes_common_scope_data_source.go
@@ -19,7 +19,7 @@ var (
 )
 
 // Create a Administrative Account data source
-func NewOauthAuthServerSettingsScopesCommonScopeDataSource() datasource.DataSource {
+func OauthAuthServerSettingsScopesCommonScopeDataSource() datasource.DataSource {
 	return &oauthAuthServerSettingsScopesCommonScopeDataSource{}
 }