+++ internal/resource/config/serversettings/logsettings/server_settings_log_settings_data_source.go
@@ -20,7 +20,7 @@ var (
 )
 
 // ServerSettingsLogSettingsDataSource is a helper function to simplify the provider implementation.
-func NewServerSettingsLogSettingsDataSource() datasource.DataSource {
+func ServerSettingsLogSettingsDataSource() datasource.DataSource {
 	return &serverSettingsLogSettingsDataSource{}
 }